--- feature-dogs/src/main/java/feature/ricky/data/remoteDataSource/RemoteDataSource.kt
@@ -1,7 +1,0 @@
-package feature.ricky.data.remoteDataSource
-
-import feature.ricky.domain.model.BreedDomain
-
-internal interface RemoteDataSource {
-    suspend fun getBreedsByPage(page: Int): List<BreedDomain>
-}
