--- feature-dogs/src/main/java/feature/ricky/domain/repository/DogsApiRepository.kt
@@ -1,7 +1,0 @@
-package feature.ricky.domain.repository
-
-import feature.ricky.presentation.model.BreedPresentation
-
-internal interface DogsApiRepository {
-    suspend fun getBreeds(page: Int): List<BreedPresentation>
-}
