+++ feature-dogs/src/main/res/layout/activity_dog_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="feature.ricky.presentation.details.DogDetailActivity">
+    tools:context="feature.dogs.presentation.details.DogDetailActivity">
 
     <TextView
         android:id="@+id/breedTitle"
