+++ feature-dogs/src/main/res/layout/activity_dogs_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="feature.ricky.presentation.home.DogsHomeActivity">
+    tools:context="feature.dogs.presentation.home.DogsHomeActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/dogsRecycler"
